requirements.txt
fastapi==0.104.1
mangum==0.17.0
Jinja2==3.1.2
openai==1.3.0
requests==2.31.0
python-dotenv==1.0.0
langchain==0.0.350
langchain-openai==0.0.5
langchain-core==0.1.0
langchain-community==0.0.12
faiss-cpu==1.7.4
numpy==1.24.3
pydantic==2.5.0
